requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2
requests==2.31.0 